refactor: Remove "core" folder from src
