+++ src/components/CarouselElement.svelte
