+++ src/lib/schema/hackathon.ts
