+++ src/routes/mailing.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Header from "../core/components/Header.svelte";
+	import Header from "../components/Header.svelte";
 	import Footer from "../components/Footer.svelte";
 
 	import { onMount } from "svelte";
